v.037 build15/11/13-15h30 [Task+NGUI]
This version contains a big reworks on the task system with a texture
over the Spartan for the state of the task. This texture use NGUI.

Task will probably be rnamed to Misson. I will implement the concept for
the Main Mission and the Secondary / Random Position. I will also redo
the GUI with NGUI and add more task while upgrading the ClaimReward()
function.